feat(posts): 게시글 수정 및 삭제에 JWT 권한 검사 적용
- 게시글 수정 API에 JWT 인증 Guard 적용
- 게시글 삭제 API에 JWT 인증 Guard 적용
- 인증 사용자와 게시글 작성자 비교 로직 구현
- 게시글 작성자 본인만 수정 및 삭제 가능하도록 구현
- 다른 사용자의 수정 및 삭제 요청에 403 Forbidden 처리
- 사용자별 게시글 수정 및 삭제 권한 테스트 완료

+++ src/modules/posts/posts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { PostEntity } from '../../entities/post.entity';
 import { Repository } from 'typeorm';
@@ -77,16 +77,31 @@ export class PostsService {
         return post;
     }
 
-    async update(id: number, updatePostDto: UpdatePostDto): Promise<PostEntity> {
+    async update(
+        id: number, 
+        updatePostDto: UpdatePostDto,
+        userId: number,
+    ): Promise<PostEntity> {
         const post = await this.findOne(id);
+
+        if(post.author.id !== userId){
+            throw new ForbiddenException("You have no rights for update this post.");
+        }
 
         Object.assign(post, updatePostDto);
 
         return this.postsRepository.save(post);
     }
 
-    async remove(id: number): Promise<void> {
+    async remove(
+        id: number,
+        userId: number,
+    ): Promise<void> {
         const post = await this.findOne(id);
+
+        if(post.author.id !== userId) {
+            throw new ForbiddenException("You have no rights for remove this post.");
+        }
 
         await this.postsRepository.remove(post);
     }
